ladder_manager/requirements.txt
# dependencies
discord.py

# dev-dependencies
pytest